# Conflicts:
#	src/DM2199/Application/Release/Application.log
#	src/DM2199/Application/Release/Application.obj
#	src/DM2199/Application/Release/Application.tlog/CL.read.1.tlog
#	src/DM2199/Application/Release/Application.tlog/CL.write.1.tlog
#	src/DM2199/Application/Release/Application.tlog/cl.command.1.tlog
#	src/DM2199/Application/Release/Application.tlog/link.read.1.tlog
#	src/DM2199/Application/Release/Bullet.obj
#	src/DM2199/Application/Release/Enemy.obj
#	src/DM2199/Application/Release/Object.obj
#	src/DM2199/Application/Release/SceneBase.obj
#	src/DM2199/Application/Release/Sound.obj
#	src/DM2199/Application/Release/vc120.pdb
#	src/DM2199/Application/Source/SceneBase.cpp
#	src/DM2199/Release/Application.exe
#	src/DM2199/Release/Application.pdb

+++ src/DM2199/Application/Source/Camera2.cpp
@@ -112,7 +112,7 @@ void Camera2::Update(double dt)
             position.y >= 0.f && position.y <= 128.f &&
             position.z >= 0.f && position.z <= 128.f)
         {
-            position += view * dt;
+            position += view * (float)dt;
         }
 
         if (position.x > 128.f)
@@ -156,7 +156,7 @@ void Camera2::Update(double dt)
             position.y >= 0.f && position.y <= 128.f &&
             position.z >= 0.f && position.z <= 128.f)
         {
-            position -= view * dt;
+            position -= view * (float)dt;
         }
 
         if (position.x > 128.f)
